remove dead code of an abandoned plan
was thinking about using progressbars to visually show resource gain
of each exped in proportion, but perhaps that's a big overhead to pay
and doesn't give us too much info. I'll maybe consider doing so in
future but for now we are cleaning up and hopefully we are not going to
lose any code thanks to version control.